# docs/api/openapi.yaml
openapi: 3.0.3
...
info:
  title: Jambo API
  version: 1.0.0
  description: |
    Catalogue, authentication, devices and playback for the Jambo Android and
    Android TV app. Introduced 2026-09-08 alongside
    `docs/plans/mobile-offline-app.md`.

    Download and offline endpoints are Phase 3 and are not described here yet.
servers:
  - url: https://jambofilms.com/api/v1
    description: production
  - url: http://localhost:8090/api/v1
    description: local
...
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      tags: [Viewer]
      summary: The viewer's notifications
      description: |
        **This is the fallback that makes push safe.** Push is an accelerator,
        never the transport — an OEM battery manager kills processes, tokens go
        stale, handsets sit offline for a day. Everything a push says is
        readable here, and the app must be usable with push switched off
        entirely.

        Carries `unread_count`, so a badge costs one request rather than two.
        The count is of the whole inbox and never of the filtered page: it is the
        bell's badge, and a badge that moved when a chip was pressed would be
        reporting the chip.

        `category` filters on the server because the list is cursor-paginated —
        filtering the page the app happens to hold would hide matching rows until
        the viewer scrolled far enough to load them. An unrecognised category is a
        422, never a silently unfiltered list.
      parameters:
        - { name: cursor, in: query, schema: { type: string } }
        - name: category
          in: query
          schema: { type: string, enum: [movies, series, account, offers] }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - properties:
                      data:
                        type: object
                        properties:
                          unread_count: { type: integer }
                          items: { type: array, items: { $ref: '#/components/schemas/Notification' } }
                          next_cursor: { type: string, nullable: true }
    delete:
      tags: [Viewer]
      summary: Clear every notification
      responses:
        '200':
          description: Cleared
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeOk' } } }
...
  /notifications/{id}:
    delete:
      tags: [Viewer]
      summary: Delete one
      description: Idempotent — deleting one already gone is a success, so a retry cannot fail.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/EnvelopeOk' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: A Sanctum personal access token from /auth/login, bound to one device.
  schemas:
    EnvelopeOk:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, enum: [true] }
        message: { type: string }
        data: { type: object }
...
    Notification:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        message: { type: string }
        icon:
          type: string
          description: |
            A Phosphor icon name, as the website's own inbox uses — `ph-film-strip`,
            `ph-device-mobile`, `ph-bell`. The app maps these onto the same glyphs.
        colour:
          type: string
          description: |
            Which tone the icon tile wears. The website paints it as
            `bg-{colour}-subtle` on `text-{colour}-emphasis`; the app captures those
            pairs so both surfaces tint a notification identically.
          enum: [primary, success, warning, danger, info]
        image_url: { type: string, nullable: true }
        action_url: { type: string, nullable: true }
        category:
          type: string
          nullable: true
          description: |
            Which filter chip this notification sits under, or null when it belongs
            under All alone. Derived from the notification's stored class, so rows
            written before the chips existed still answer it.
          enum: [movies, series, account, offers]
        read: { type: boolean }
        created_at: { type: string, format: date-time, nullable: true }
...